moved tests to namespace which reflects the source code
