+++ src/test/java/com/wb3tech/kernel/entity/EntityTests.java
@@ -1,7 +1,5 @@
-package com.wb3tech.kernel;
+package com.wb3tech.kernel.entity;
 
-import com.wb3tech.kernel.entity.Entity;
-import com.wb3tech.kernel.entity.Identity;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
